palindrome.asm: An 8086 assembly program.
When run the program will ask the user to enter a word. Once inputted
the program will check the to see if the word is a palindrome. The
results will be printed to screen.

Uses three separate procedures: get_length, convert_lower_, and
palindrome.

Get_length will get the length of the inputted string by checking for
the first non ascii 0 or reaching 50 characters in length(the set memory
length for the string).

convert_lower will convert any uppercase to lower case by checking the
hex value of the character and adding 32 if beneath the threshold.

palindrome will compare characters at each end of the string and returns
0 (true) if the inputted string is a palindrome.