Completed BubbleSort in assembly 8086.
The program will ask the user for 5 numbers which will be stored in an
array. Once the 5th number is stored, the program will display the
unsorted array before sorting. Once sorted, the program will display the
array one last time.

print_msg: Takes a string and prints it to screen using the print_string
procedure from the emu8086 library.

get_numbers: Takes the length as a param. Uses a for loop to get the numbers from the user. User numbers are initialy stored in the cx register after input. The program stores it in the array and resets the cx register.

print_arr: Prints the array content to the console. Uses the print_num
procedure from the emu8086 library to print individual numbers. Uses a
for loop to go through the entire array.

bubble_sort: Uses the length of the array as the parameter. Makes
multiple pass through the array. With each pass through the array, swap
the cells if the current cell is larger then the next one. Stops if no
swap was made on the last iteration.